fix: app crash
