+++ packages/plugins/Trader/src/base.ts
@@ -2,7 +2,7 @@ import type { Plugin } from '@masknet/plugin-infra'
 import { DEFAULT_PLUGIN_PUBLISHER, EnhanceableSite } from '@masknet/shared-base'
 import { PLUGIN_ID } from './constants/index.js'
 import { languages } from './locales/languages.js'
-import { TRADER_WEB3_CONFIG } from '@masknet/shared'
+import { TRADER_WEB3_CONFIG } from './config.js'
 
 export const base: Plugin.Shared.Definition = {
     ID: PLUGIN_ID,
