+++ packages/plugins/Trader/src/index.ts
@@ -2,3 +2,4 @@ export * from './messages.js'
 export * from './trader/useAllProviderTradeContext.js'
 export * from './SiteAdaptor/trader/Trader.js'
 export * from './locales/languages.js'
+export * from './config.js'
